Refactor - Extract out Haproxy helpers
