+++ lib/litmus_paper.rb
@@ -22,6 +22,7 @@ require 'litmus_paper/dependency/haproxy_backends'
 require 'litmus_paper/dependency/http'
 require 'litmus_paper/dependency/script'
 require 'litmus_paper/dependency/tcp'
+require 'litmus_paper/haproxy_util'
 require 'litmus_paper/health'
 require 'litmus_paper/logger'
 require 'litmus_paper/metric/big_brother_service'
